requirements.txt
selenium
requests
python-dotenv
browserstack-sdk